+++ pulse-binding/src/mainloop/events/timer.rs
@@ -31,21 +31,22 @@ pub struct TimeEvent<T>
     ptr: *mut TimeEventInternal,
     /// Source mainloop
     owner: Rc<T>,
+    /// Saved callback closure, for later destruction
+    _saved_cb: EventCb,
 }
 
-/// A time event callback prototype
-pub type TimeEventCb = extern "C" fn(a: *const MainloopApi, e: *mut TimeEventInternal,
-    tv: *const timeval, userdata: *mut c_void);
-/// A time event destroy callback prototype
-pub type TimeEventDestroyCb = extern "C" fn(a: *const MainloopApi, e: *mut TimeEventInternal,
-    userdata: *mut c_void);
+pub(crate) type EventCb = ::callbacks::MultiUseCallback<FnMut(),
+    extern "C" fn(a: *const MainloopApi, e: *mut TimeEventInternal, tv: *const timeval,
+    userdata: *mut c_void)>;
 
 impl<T> TimeEvent<T>
     where T: MainloopInnerType
 {
-    pub(crate) fn from_raw(ptr: *mut TimeEventInternal, mainloop_inner: Rc<T>) -> Self {
+    pub(crate) fn from_raw(ptr: *mut TimeEventInternal, mainloop_inner: Rc<T>, callback: EventCb
+        ) -> Self
+    {
         assert_eq!(false, ptr.is_null());
-        Self { ptr: ptr, owner: mainloop_inner }
+        Self { ptr: ptr, owner: mainloop_inner, _saved_cb: callback }
     }
 
     /// Needed to support
@@ -60,13 +61,6 @@ impl<T> TimeEvent<T>
         let fn_ptr = (*self.owner).get_api().time_restart.unwrap();
         fn_ptr(self.ptr, &tv.0);
     }
-
-    /// Set a function that is called when the timer event source is destroyed.
-    /// Use this to free the userdata argument if required.
-    pub fn set_destroy(&mut self, cb: TimeEventDestroyCb) {
-        let fn_ptr = (*self.owner).get_api().time_set_destroy.unwrap();
-        fn_ptr(self.ptr, Some(cb));
-    }
 }
 
 impl<T> Drop for TimeEvent<T>
@@ -76,4 +70,16 @@ impl<T> Drop for TimeEvent<T>
         let fn_ptr = (*self.owner).get_api().time_free.unwrap();
         fn_ptr(self.ptr);
     }
+}
+
+/// Proxy for the event callback.
+/// Warning: This is for multi-use cases! It does **not** destroy the actual closure callback, which
+/// must be accomplished separately to avoid a memory leak.
+pub(crate)
+extern "C"
+fn event_cb_proxy(_: *const MainloopApi, _: *mut TimeEventInternal, _: *const timeval,
+    userdata: *mut c_void)
+{
+    let callback = EventCb::get_callback(userdata);
+    callback();
 }